github/webhook: refactor
