+++ backend/server/src/github/webhook/mod.rs
@@ -1,79 +1,27 @@
+// GitHub webhook event handling
+
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use anyhow::{Context, bail};
 use octocrab::Octocrab;
-use octocrab::models::pulls::PullRequest;
-use octocrab::models::webhook_events::{EventInstallation, WebhookEventPayload as WEP, payload};
+use octocrab::models::webhook_events::{EventInstallation, WebhookEventPayload as WEP};
 use serde::Deserialize;
 use tokio::sync::mpsc;
-use tracing::{debug, warn};
 
 use crate::config::GitHubAppConfig;
 use crate::db::DbService;
 use crate::git::GitTask;
-use crate::github::{CICheckInfo, GitHubTask};
-use crate::github_permissions::{PermissionContext, check_github_app_permission};
+use crate::github::GitHubTask;
 
 pub(crate) mod comment_command;
+mod comments;
+mod installations;
+mod merge_queue;
+mod pull_requests;
+mod reviews;
+mod workflows;
 
-pub async fn handle_webhook_payload(
-    webhook_payload: WEP,
-    repository_info: Option<(String, String)>, // (owner, repo_name)
-    installation: Option<EventInstallation>,
-    git_sender: mpsc::Sender<GitTask>,
-    github_sender: Option<mpsc::Sender<GitHubTask>>,
-    octocrab: Option<Octocrab>,
-    require_approval: bool,
-    merge_queue_require_approval: bool,
-    db_service: DbService,
-    github_app_configs: Arc<HashMap<String, GitHubAppConfig>>,
-) {
-    match webhook_payload {
-        WEP::PullRequest(pr) => {
-            handle_github_pr(
-                *pr,
-                git_sender,
-                github_sender,
-                require_approval,
-                db_service,
-                github_app_configs,
-            )
-            .await
-        },
-        WEP::WorkflowRun(workflow_run) => {
-            handle_github_workflow_run(*workflow_run, git_sender, octocrab).await
-        },
-        WEP::MergeGroup(merge_group) => {
-            handle_github_merge_group(
-                *merge_group,
-                repository_info,
-                git_sender,
-                github_sender,
-                merge_queue_require_approval,
-                db_service.clone(),
-                github_app_configs,
-            )
-            .await
-        },
-        WEP::Installation(installation_payload) => {
-            handle_github_installation(*installation_payload, installation, db_service).await
-        },
-        WEP::InstallationRepositories(installation_repos) => {
-            handle_github_installation_repositories(*installation_repos, installation, db_service)
-                .await
-        },
-        WEP::PullRequestReview(review) => {
-            handle_github_pr_review(*review, github_sender, db_service).await
-        },
-        WEP::IssueComment(comment_event) => {
-            handle_github_issue_comment(*comment_event, repository_info, github_sender).await
-        },
-        // We probably don't want to react to every push
-        // WEP::Push(pr) => handle_github_push(*pr).await,
-        _ => (),
-    }
-}
+// Webhook payload data structures
 
 #[derive(Debug, Deserialize)]
 struct WorkflowRunPullRequest {
@@ -124,1172 +72,72 @@ struct MergeGroupUser {
     email: String,
 }
 
-async fn is_valid_user(pr: &PullRequest, db_service: &DbService) -> anyhow::Result<()> {
-    let pr_author = &pr.user.as_ref().context("Missing github user for PR")?;
-    let username = &pr_author.login;
-    let user_id = pr_author.id.0 as i64;
+// Main webhook dispatcher
 
-    if !db_service.is_user_approved(username, user_id).await? {
-        bail!("User is not approved for workflows");
-    }
-
-    Ok(())
-}
-
-/// Store PR metadata in the database
-async fn store_pr_metadata(pr: &PullRequest, state: &str, db_service: &DbService) {
-    let pr_author = match &pr.user {
-        Some(user) => &user.login,
-        None => {
-            warn!(
-                "PR #{} has no user information, skipping metadata storage",
-                pr.number
-            );
-            return;
-        },
-    };
-
-    let pr_number = pr.number as i64;
-
-    let owner = match &pr.base.repo {
-        Some(repo) => match &repo.owner {
-            Some(owner) => &owner.login,
-            None => {
-                warn!("PR base repo has no owner");
-                return;
-            },
-        },
-        None => {
-            warn!("PR base has no repo");
-            return;
-        },
-    };
-
-    let repo_name = match &pr.base.repo {
-        Some(repo) => &repo.name,
-        None => {
-            warn!("PR base has no repo");
-            return;
-        },
-    };
-
-    let head_sha = &pr.head.sha;
-    let base_sha = &pr.base.sha;
-
-    let title = match &pr.title {
-        Some(t) => t.as_str(),
-        None => "",
-    };
-
-    let created_at = pr.created_at.map(|dt| dt.to_rfc3339()).unwrap_or_default();
-    let updated_at = pr.updated_at.map(|dt| dt.to_rfc3339()).unwrap_or_default();
-
-    if let Err(e) = db_service
-        .upsert_pull_request(
-            pr_number,
-            owner,
-            repo_name,
-            head_sha,
-            base_sha,
-            title,
-            pr_author,
-            state,
-            &created_at,
-            &updated_at,
-        )
-        .await
-    {
-        warn!("Failed to store PR metadata for PR #{}: {:?}", pr_number, e);
-    } else {
-        debug!("Stored metadata for PR #{}", pr_number);
-    }
-}
-
-/// On `Synchronize`, cancel any pending `@eka-ci merge` whose pinned SHA
-/// no longer matches the head. Posts a notification, reacts `:confused:`
-/// on the original command, and clears the DB fields. Per-step failures
-/// are logged but non-fatal.
-async fn check_comment_merge_drift(
-    pr: &PullRequest,
-    db_service: &DbService,
-    github_sender: &mpsc::Sender<GitHubTask>,
-) {
-    let Some(base_repo) = pr.base.repo.as_ref() else {
-        return;
-    };
-    let Some(owner) = base_repo.owner.as_ref().map(|o| o.login.clone()) else {
-        return;
-    };
-    let repo_name = base_repo.name.clone();
-    let pr_number = pr.number as i64;
-
-    let pr_row = match crate::db::github::get_pull_request_row(
-        &owner,
-        &repo_name,
-        pr_number,
-        &db_service.pool,
-    )
-    .await
-    {
-        Ok(Some(row)) => row,
-        Ok(None) => return,
-        Err(e) => {
-            warn!(
-                "Failed to look up PR row for drift check on {}/{}#{}: {:?}",
-                owner, repo_name, pr_number, e
-            );
-            return;
-        },
-    };
-
-    let Some(pending) = pr_row.pending_comment_merge() else {
-        return; // nothing pending
-    };
-
-    if pending.sha == pr.head.sha {
-        // Possible on a no-op force-push; just skip.
-        return;
-    }
-
-    debug!(
-        "Comment-merge SHA drift on {}/{}#{}: pinned={} new_head={}",
-        owner, repo_name, pr_number, pending.sha, pr.head.sha
-    );
-
-    if let Err(e) = github_sender
-        .send(GitHubTask::CommentMergeDriftCancelled {
-            owner: owner.clone(),
-            repo_name: repo_name.clone(),
-            pr_number,
-            expected_sha: pending.sha.clone(),
-            actual_sha: pr.head.sha.clone(),
-            requester_login: pending.requester_login.clone(),
-        })
-        .await
-    {
-        warn!(
-            "Failed to enqueue CommentMergeDriftCancelled for {}/{}#{}: {:?}",
-            owner, repo_name, pr_number, e
-        );
-    }
-
-    // `:confused:` the original command so the thread reflects the cancel.
-    if let Err(e) = github_sender
-        .send(GitHubTask::ReactToComment {
-            owner: owner.clone(),
-            repo_name: repo_name.clone(),
-            comment_id: pending.comment_id,
-            content: "confused",
-        })
-        .await
-    {
-        warn!(
-            "Failed to enqueue drift-cancel reaction on {}/{}#{}: {:?}",
-            owner, repo_name, pr_number, e
-        );
-    }
-
-    // Clear pending so downstream schedulers don't re-fire on the drifted head.
-    if let Err(e) = crate::db::github::clear_comment_merge_request(
-        &owner,
-        &repo_name,
-        pr_number,
-        &db_service.pool,
-    )
-    .await
-    {
-        warn!(
-            "Failed to clear comment_merge_request for {}/{}#{}: {:?}",
-            owner, repo_name, pr_number, e
-        );
-    }
-}
-
-async fn handle_github_pr(
-    pr: payload::PullRequestWebhookEventPayload,
+pub async fn handle_webhook_payload(
+    webhook_payload: WEP,
+    repository_info: Option<(String, String)>, // (owner, repo_name)
+    installation: Option<EventInstallation>,
     git_sender: mpsc::Sender<GitTask>,
     github_sender: Option<mpsc::Sender<GitHubTask>>,
+    octocrab: Option<Octocrab>,
     require_approval: bool,
-    db_service: DbService,
-    github_app_configs: Arc<HashMap<String, GitHubAppConfig>>,
-) {
-    use payload::PullRequestWebhookEventAction as PRWEA;
-
-    // This handler should only exist if a valid github_sender also exsits,
-    // should be safe to assume this will always succeed
-    let github_sender = match github_sender {
-        Some(sender) => sender,
-        _ => {
-            warn!("GitHub service is down, unable to service webhook request. Restart Eka-CI");
-            return;
-        },
-    };
-
-    match pr.action {
-        PRWEA::Opened | PRWEA::Synchronize | PRWEA::Reopened => {
-            debug!("Received event for PR #{}", &pr.pull_request.number);
-
-            // Check GitHub App permissions before processing
-            if let Some(base_repo) = &pr.pull_request.base.repo {
-                let owner = &base_repo
-                    .owner
-                    .as_ref()
-                    .map(|o| o.login.clone())
-                    .unwrap_or_default();
-                let repo_name = &base_repo.name;
-                let branch = &pr.pull_request.base.ref_field;
-
-                let permission_context = PermissionContext {
-                    repo_owner: owner.clone(),
-                    repo_name: repo_name.clone(),
-                    branch: Some(branch.clone()),
-                };
-
-                // Check if any configured GitHub App has permission for this repo/branch
-                let has_permission = github_app_configs
-                    .values()
-                    .any(|config| check_github_app_permission(config, &permission_context).is_ok());
-
-                if !has_permission {
-                    warn!(
-                        "No GitHub App has permission for repository {}/{} branch {}. Skipping PR \
-                         #{}",
-                        owner, repo_name, branch, pr.pull_request.number
-                    );
-                    return;
-                }
-            }
-
-            // Store PR metadata in database
-            store_pr_metadata(&pr.pull_request, "open", &db_service).await;
-
-            // SHA-drift cancellation for pending @eka-ci merge commands.
-            // `@eka-ci merge` is SHA-pinned: once a user issues the command,
-            // we refuse to merge a newer head (new commits imply unreviewed
-            // changes). On Synchronize, if the stored pinned sha differs
-            // from the new head, clear the pending request and notify the
-            // requester. Opened/Reopened can't have drifted state worth
-            // cancelling (no prior comment-merge on a just-opened PR; a
-            // reopened PR's previous pending merge would already have been
-            // cleared at close-time).
-            if matches!(pr.action, PRWEA::Synchronize) {
-                check_comment_merge_drift(&pr.pull_request, &db_service, &github_sender).await;
-            }
-
-            if require_approval && is_valid_user(&pr.pull_request, &db_service).await.is_err() {
-                let username = match &pr.pull_request.user {
-                    Some(user) => user.login.clone(),
-                    None => {
-                        warn!(
-                            "PR #{} has no user information, skipping approval gate",
-                            pr.pull_request.number
-                        );
-                        return;
-                    },
-                };
-                let ci_check_info = match CICheckInfo::from_gh_pr_head(&pr.pull_request) {
-                    Ok(info) => info,
-                    Err(e) => {
-                        warn!(
-                            "PR #{} missing head repo/owner data, cannot create approval check: \
-                             {:?}",
-                            pr.pull_request.number, e
-                        );
-                        return;
-                    },
-                };
-                if let Err(e) = github_sender
-                    .send(GitHubTask::CreateApprovalRequiredCheckRun {
-                        ci_check_info: std::sync::Arc::new(ci_check_info),
-                        username,
-                    })
-                    .await
-                {
-                    warn!("Failed to send approval required check run task: {:?}", e);
-                }
-
-                // Don't proceed until workflow is approved
-                return;
-            }
-
-            let git_task = GitTask::GitHubCheckout(pr.pull_request.clone());
-
-            if let Err(e) = git_sender.send(git_task).await {
-                warn!("Failed to send PR checkout task: {:?}", e);
-            } else {
-                debug!(
-                    "Successfully queued PR checkout task for PR #{}",
-                    pr.pull_request.number
-                );
-            }
-        },
-        PRWEA::Closed | PRWEA::ConvertedToDraft => {
-            debug!(
-                "Received close/draft event for PR #{}, cancelling check runs",
-                pr.pull_request.number
-            );
-
-            // Update PR state in database
-            let state = if matches!(pr.action, PRWEA::Closed) {
-                // Determine if merged or just closed
-                if pr.pull_request.merged_at.is_some() {
-                    "merged"
-                } else {
-                    "closed"
-                }
-            } else {
-                "open" // ConvertedToDraft keeps it open
-            };
-            store_pr_metadata(&pr.pull_request, state, &db_service).await;
-
-            let ci_check_info = match CICheckInfo::from_gh_pr_head(&pr.pull_request) {
-                Ok(info) => info,
-                Err(e) => {
-                    warn!(
-                        "PR #{} missing head repo/owner data, cannot cancel check runs: {:?}",
-                        pr.pull_request.number, e
-                    );
-                    return;
-                },
-            };
-
-            if let Err(e) = github_sender
-                .send(GitHubTask::CancelCheckRunsForCommit {
-                    ci_check_info: std::sync::Arc::new(ci_check_info),
-                })
-                .await
-            {
-                warn!(
-                    "Failed to send cancellation task for PR #{}: {:?}",
-                    pr.pull_request.number, e
-                );
-            } else {
-                debug!(
-                    "Successfully queued cancellation task for PR #{}",
-                    pr.pull_request.number
-                );
-            }
-        },
-        PRWEA::Enqueued => {
-            debug!(
-                "Received enqueued event for PR #{}",
-                &pr.pull_request.number
-            );
-
-            // Store PR metadata in database
-            store_pr_metadata(&pr.pull_request, "open", &db_service).await;
-
-            if require_approval && is_valid_user(&pr.pull_request, &db_service).await.is_err() {
-                let username = match &pr.pull_request.user {
-                    Some(user) => user.login.clone(),
-                    None => {
-                        warn!(
-                            "PR #{} has no user information, skipping approval gate",
-                            pr.pull_request.number
-                        );
-                        return;
-                    },
-                };
-                let ci_check_info = match CICheckInfo::from_gh_pr_head(&pr.pull_request) {
-                    Ok(info) => info,
-                    Err(e) => {
-                        warn!(
-                            "PR #{} missing head repo/owner data, cannot create approval check: \
-                             {:?}",
-                            pr.pull_request.number, e
-                        );
-                        return;
-                    },
-                };
-                if let Err(e) = github_sender
-                    .send(GitHubTask::CreateApprovalRequiredCheckRun {
-                        ci_check_info: std::sync::Arc::new(ci_check_info),
-                        username,
-                    })
-                    .await
-                {
-                    warn!("Failed to send approval required check run task: {:?}", e);
-                }
-
-                // Don't proceed until workflow is approved
-                return;
-            }
-
-            let git_task = GitTask::GitHubCheckout(pr.pull_request.clone());
-
-            if let Err(e) = git_sender.send(git_task).await {
-                warn!("Failed to send PR checkout task for enqueued PR: {:?}", e);
-            } else {
-                debug!(
-                    "Successfully queued PR checkout task for enqueued PR #{}",
-                    pr.pull_request.number
-                );
-            }
-        },
-        action => {
-            debug!("Ignoring non-actionable PR action: {:?}", &action);
-        },
-    }
-}
-
-async fn handle_github_workflow_requested(
-    payload: WorkflowRunData,
-    git_sender: mpsc::Sender<GitTask>,
-    octocrab: Option<Octocrab>,
-) -> anyhow::Result<()> {
-    // Check if there are any pull requests associated with this workflow run
-    if payload.pull_requests.is_empty() {
-        bail!("No pull requests associated with this workflow_run (likely from a fork or push)");
-    }
-
-    let owner = payload.repository.owner.login;
-    let repo_name = payload.repository.name;
-
-    // Fetch the full PR details using octocrab
-    let octocrab = octocrab.context("GitHub App octocrab instance not available")?;
-
-    for pr_number in payload.pull_requests {
-        debug!(
-            "Workflow approved for PR #{} in {}/{}",
-            &pr_number.number, &owner, &repo_name
-        );
-        let pull_request = octocrab
-            .pulls(&owner, &repo_name)
-            .get(pr_number.number)
-            .await?;
-        // Workflow was approved by maintainer, send the PR for checkout
-        let git_task = GitTask::GitHubCheckout(pull_request);
-        git_sender.send(git_task).await?;
-    }
-
-    Ok(())
-}
-
-async fn handle_github_workflow_run(
-    workflow_run: payload::WorkflowRunWebhookEventPayload,
-    git_sender: mpsc::Sender<GitTask>,
-    octocrab: Option<Octocrab>,
-) {
-    use payload::WorkflowRunWebhookEventAction as WRWEA;
-
-    // Only handle "requested" action (when workflow is approved by maintainer)
-    if workflow_run.action != WRWEA::Requested {
-        debug!("Ignoring workflow_run action: {:?}", workflow_run.action);
-        return;
-    }
-
-    debug!("Received workflow_run requested event");
-
-    // Parse the workflow_run data to extract PR information
-    let workflow_run_data: WorkflowRunData = match serde_json::from_value(workflow_run.workflow_run)
-    {
-        Ok(data) => data,
-        Err(e) => {
-            warn!("Failed to parse workflow_run data: {:?}", e);
-            return;
-        },
-    };
-
-    if let Err(e) = handle_github_workflow_requested(workflow_run_data, git_sender, octocrab).await
-    {
-        warn!("Failed to process github workflow requested: {}", e);
-    }
-}
-
-async fn handle_github_merge_group(
-    merge_group_payload: payload::MergeGroupWebhookEventPayload,
-    repository_info: Option<(String, String)>,
-    git_sender: mpsc::Sender<GitTask>,
-    github_sender: Option<mpsc::Sender<GitHubTask>>,
     merge_queue_require_approval: bool,
     db_service: DbService,
     github_app_configs: Arc<HashMap<String, GitHubAppConfig>>,
 ) {
-    use payload::MergeGroupWebhookEventAction as MGWEA;
-
-    // This handler should only exist if a valid github_sender also exists
-    let _github_sender = match github_sender {
-        Some(sender) => sender,
-        _ => {
-            warn!("GitHub service is down, unable to service merge_group webhook. Restart Eka-CI");
-            return;
+    match webhook_payload {
+        WEP::PullRequest(pr) => {
+            pull_requests::handle_github_pr(
+                *pr,
+                git_sender,
+                github_sender,
+                require_approval,
+                db_service,
+                github_app_configs,
+            )
+            .await
         },
-    };
-
-    // Only handle "checks_requested" action
-    if merge_group_payload.action != MGWEA::ChecksRequested {
-        debug!(
-            "Ignoring merge_group action: {:?}",
-            merge_group_payload.action
-        );
-        return;
-    }
-
-    debug!("Received merge_group checks_requested event");
-
-    // Extract repository info
-    let (owner, repo_name) = match repository_info {
-        Some(info) => info,
-        None => {
-            warn!("Missing repository info in merge_group webhook");
-            return;
+        WEP::WorkflowRun(workflow_run) => {
+            workflows::handle_github_workflow_run(*workflow_run, git_sender, octocrab).await
         },
-    };
-
-    // Parse the merge_group data
-    let merge_group: MergeGroupData = match serde_json::from_value(merge_group_payload.merge_group)
-    {
-        Ok(data) => data,
-        Err(e) => {
-            warn!("Failed to parse merge_group data: {:?}", e);
-            return;
+        WEP::MergeGroup(merge_group) => {
+            merge_queue::handle_github_merge_group(
+                *merge_group,
+                repository_info,
+                git_sender,
+                github_sender,
+                merge_queue_require_approval,
+                db_service.clone(),
+                github_app_configs,
+            )
+            .await
         },
-    };
-
-    debug!(
-        "Processing merge queue commit {} for ref {} in {}/{}",
-        merge_group.head_sha, merge_group.head_ref, owner, repo_name
-    );
-
-    // Check GitHub App permissions before processing merge group
-    let permission_context = PermissionContext {
-        repo_owner: owner.clone(),
-        repo_name: repo_name.clone(),
-        branch: Some(merge_group.base_ref.clone()),
-    };
-
-    // Check if any configured GitHub App has permission for this repo/branch
-    let has_permission = github_app_configs
-        .values()
-        .any(|config| check_github_app_permission(config, &permission_context).is_ok());
-
-    if !has_permission {
-        warn!(
-            "No GitHub App has permission for repository {}/{} branch {}. Skipping merge group \
-             for commit {}",
-            owner, repo_name, merge_group.base_ref, merge_group.head_sha
-        );
-        return;
-    }
-
-    // Check if approval is required for merge queue builds
-    if merge_queue_require_approval {
-        let author = &merge_group.head_commit.author.name;
-
-        // Check if the author is approved
-        // Note: Merge groups don't provide user_id, so we use 0 as a sentinel
-        // The approval check will only check by username in this case
-        match db_service.is_user_approved(author, 0).await {
-            Ok(true) => {
-                debug!("Merge queue build approved for user: {}", author);
-            },
-            Ok(false) => {
-                warn!(
-                    "Merge queue build requires approval. User {} is not approved for {}/{}",
-                    author, owner, repo_name
-                );
-                return;
-            },
-            Err(e) => {
-                warn!(
-                    "Failed to check approval status for user {}: {:?}",
-                    author, e
-                );
-                return;
-            },
-        }
-    }
-
-    // Store merge queue build in database
-    let now = chrono::Utc::now().to_rfc3339();
-    if let Err(e) = db_service
-        .upsert_merge_queue_build(
-            &owner,
-            &repo_name,
-            &merge_group.head_sha,
-            &merge_group.base_sha,
-            &merge_group.base_ref,
-            &merge_group.head_sha, // merge_group_head_sha is same as head_sha
-            &now,
-            &now,
-        )
-        .await
-    {
-        warn!(
-            "Failed to store merge queue build for commit {}: {:?}",
-            merge_group.head_sha, e
-        );
-        // Continue processing even if DB storage fails
-    }
-
-    // Create a synthetic PullRequest to reuse the existing GitTask::GitHubCheckout flow
-    // This allows us to leverage all the existing git checkout and CI evaluation logic
-    use octocrab::models::Repository as OctoRepo;
-    use octocrab::models::pulls::{Head, PullRequest};
-
-    // Create minimal repository object
-    let repo_value = serde_json::json!({
-        "name": repo_name,
-        "owner": {
-            "login": owner
+        WEP::Installation(installation_payload) => {
+            installations::handle_github_installation(
+                *installation_payload,
+                installation,
+                db_service,
+            )
+            .await
         },
-        "clone_url": format!("https://github.com/{}/{}.git", owner, repo_name),
-    });
-    let repo: OctoRepo = match serde_json::from_value(repo_value) {
-        Ok(r) => r,
-        Err(e) => {
-            warn!("Failed to create repository object: {:?}", e);
-            return;
+        WEP::InstallationRepositories(installation_repos) => {
+            installations::handle_github_installation_repositories(
+                *installation_repos,
+                installation,
+                db_service,
+            )
+            .await
         },
-    };
-
-    // Create head and base objects for the merge queue commit
-    let head_value = serde_json::json!({
-        "ref": merge_group.head_ref,
-        "sha": merge_group.head_sha,
-        "repo": repo.clone(),
-    });
-    let head: Head = match serde_json::from_value(head_value) {
-        Ok(h) => h,
-        Err(e) => {
-            warn!("Failed to create head object: {:?}", e);
-            return;
+        WEP::PullRequestReview(review) => {
+            reviews::handle_github_pr_review(*review, github_sender, db_service).await
         },
-    };
-
-    let base_value = serde_json::json!({
-        "ref": merge_group.base_ref,
-        "sha": merge_group.base_sha,
-        "repo": repo.clone(),
-    });
-    let base: Head = match serde_json::from_value(base_value) {
-        Ok(b) => b,
-        Err(e) => {
-            warn!("Failed to create base object: {:?}", e);
-            return;
-        },
-    };
-
-    // Create a synthetic PullRequest object
-    let pr_value = serde_json::json!({
-        "number": 0, // Merge queue doesn't have a PR number in this context
-        "state": "open",
-        "title": format!("Merge queue for {}", merge_group.base_ref),
-        "head": head,
-        "base": base,
-        "user": {
-            "login": merge_group.head_commit.author.name,
-            "id": 0,
-        },
-    });
-
-    let pr: PullRequest = match serde_json::from_value(pr_value) {
-        Ok(p) => p,
-        Err(e) => {
-            warn!("Failed to create synthetic PullRequest: {:?}", e);
-            return;
-        },
-    };
-
-    // Send the checkout task using the existing flow
-    let git_task = GitTask::GitHubCheckout(pr);
-
-    if let Err(e) = git_sender.send(git_task).await {
-        warn!("Failed to send merge queue checkout task: {:?}", e);
-    } else {
-        debug!(
-            "Successfully queued merge queue checkout task for commit {}",
-            merge_group.head_sha
-        );
-    }
-}
-
-/// Handle `pull_request_review` webhook events.
-///
-/// When a reviewer submits an approval (or dismisses/changes an existing review),
-/// re-check auto-merge eligibility for the PR. This ensures approvals that arrive
-/// after builds have completed still trigger the merge flow, rather than waiting
-/// for a subsequent build to reach the recorder's end-of-jobset hook.
-async fn handle_github_pr_review(
-    review_event: payload::PullRequestReviewWebhookEventPayload,
-    github_sender: Option<mpsc::Sender<GitHubTask>>,
-    db_service: DbService,
-) {
-    use octocrab::models::pulls::ReviewState;
-    use payload::PullRequestReviewWebhookEventAction as PRRWEA;
-
-    let Some(github_sender) = github_sender else {
-        warn!("GitHub service is down, unable to service review webhook. Restart Eka-CI");
-        return;
-    };
-
-    // Only react to events that actually change approval state. Edits to a
-    // review body do not change whether the PR is approved.
-    match review_event.action {
-        PRRWEA::Submitted => {
-            // Only approving reviews affect merge eligibility in a positive
-            // direction. Non-approving reviews (Commented, ChangesRequested)
-            // would only ever prevent a merge, which the existing approval
-            // check already handles at merge time, so we skip re-triggering.
-            if review_event.review.state != Some(ReviewState::Approved) {
-                debug!(
-                    "Ignoring non-approving review on PR #{}",
-                    review_event.pull_request.number
-                );
-                return;
-            }
-        },
-        PRRWEA::Dismissed => {
-            // A dismissal can affect later recomputation; re-run the check so
-            // any cached state is refreshed.
-        },
-        action => {
-            debug!("Ignoring pull_request_review action: {:?}", action);
-            return;
-        },
-    }
-
-    let pr = &review_event.pull_request;
-    let pr_number = pr.number as i64;
-
-    let Some(base_repo) = pr.base.repo.as_ref() else {
-        warn!(
-            "pull_request_review for PR #{} missing base repo, cannot check auto-merge",
-            pr_number
-        );
-        return;
-    };
-
-    let Some(owner) = base_repo.owner.as_ref().map(|o| o.login.clone()) else {
-        warn!(
-            "pull_request_review for PR #{} missing base repo owner, cannot check auto-merge",
-            pr_number
-        );
-        return;
-    };
-    let repo_name = base_repo.name.clone();
-
-    // Only consider PRs where auto-merge has been enabled and which are still
-    // open. This matches the gate applied by the scheduler-driven path.
-    let pr_record = match crate::db::github::get_pr_by_head_sha(
-        &pr.head.sha,
-        &owner,
-        &repo_name,
-        &db_service.pool,
-    )
-    .await
-    {
-        Ok(Some(record)) => record,
-        Ok(None) => {
-            debug!(
-                "No stored PR record for {}/{}#{} (head {}); skipping review-triggered auto-merge",
-                owner, repo_name, pr_number, pr.head.sha
-            );
-            return;
-        },
-        Err(e) => {
-            warn!(
-                "Failed to look up PR record for {}/{}#{}: {:?}",
-                owner, repo_name, pr_number, e
-            );
-            return;
-        },
-    };
-
-    if !pr_record.auto_merge_enabled || pr_record.state != "open" {
-        debug!(
-            "PR #{} auto-merge not enabled or not open (enabled={}, state={}); skipping",
-            pr_number, pr_record.auto_merge_enabled, pr_record.state
-        );
-        return;
-    }
-
-    let task = GitHubTask::CheckAutoMerge {
-        owner: owner.clone(),
-        repo_name: repo_name.clone(),
-        pr_number,
-    };
-
-    if let Err(e) = github_sender.send(task).await {
-        warn!(
-            "Failed to send CheckAutoMerge for PR #{} after review: {:?}",
-            pr_number, e
-        );
-    } else {
-        debug!(
-            "Queued CheckAutoMerge for PR #{} in {}/{} after review event",
-            pr_number, owner, repo_name
-        );
-    }
-}
-
-/// Handle `issue_comment` webhook events.
-///
-/// GitHub fires this for both plain-issue and PR comments (PRs are issues
-/// underneath). We filter to PR comments only, parse the body for an
-/// `@eka-ci …` command (via [`comment_command::parse_comment_command`]),
-/// and enqueue a `ProcessMergeCommand` for the GitHub service to authorize
-/// and action.
-///
-/// Comments authored by bots (including eka-ci itself) are ignored to
-/// prevent feedback loops from our own rocket/+1 reaction flow if we ever
-/// start posting parseable bodies.
-async fn handle_github_issue_comment(
-    comment_event: payload::IssueCommentWebhookEventPayload,
-    repository_info: Option<(String, String)>,
-    github_sender: Option<mpsc::Sender<GitHubTask>>,
-) {
-    use payload::IssueCommentWebhookEventAction as ICWEA;
-
-    let Some(github_sender) = github_sender else {
-        warn!("GitHub service is down, unable to service issue_comment webhook. Restart Eka-CI");
-        return;
-    };
-
-    // Only newly-created comments trigger commands; edit/delete is a no-op.
-    if comment_event.action != ICWEA::Created {
-        debug!("Ignoring issue_comment action: {:?}", comment_event.action);
-        return;
-    }
-
-    // `issue.pull_request` is `Some` iff the issue is a PR.
-    if comment_event.issue.pull_request.is_none() {
-        debug!(
-            "Ignoring comment on non-PR issue #{}",
-            comment_event.issue.number
-        );
-        return;
-    }
-
-    // Skip bot-authored comments to prevent self-reaction loops.
-    if comment_event
-        .comment
-        .user
-        .r#type
-        .eq_ignore_ascii_case("Bot")
-    {
-        debug!(
-            "Ignoring bot-authored comment on PR #{}",
-            comment_event.issue.number
-        );
-        return;
-    }
-
-    let Some(body) = comment_event.comment.body.as_ref() else {
-        debug!(
-            "Ignoring empty-body comment on PR #{}",
-            comment_event.issue.number
-        );
-        return;
-    };
-
-    // Pre-filter: skip the task queue for the vast majority (non-commands).
-    if comment_command::parse_comment_command(body).is_none() {
-        return;
-    }
-
-    let Some((owner, repo_name)) = repository_info else {
-        warn!(
-            "issue_comment on PR #{} missing repository info; cannot process command",
-            comment_event.issue.number
-        );
-        return;
-    };
-
-    // Per-(user, owner, repo) rate limit. Silent drop on rejection —
-    // any feedback would amplify the spam we're containing.
-    let requester_id = comment_event.comment.user.id.0 as i64;
-    if !comment_command::check_and_record_rate_limit(requester_id, &owner, &repo_name) {
-        debug!(
-            "Rate-limited @eka-ci command from user {} on {}/{}#{} (comment {})",
-            comment_event.comment.user.login,
-            owner,
-            repo_name,
-            comment_event.issue.number,
-            comment_event.comment.id.0
-        );
-        return;
-    }
-
-    let task = GitHubTask::ProcessMergeCommand {
-        owner,
-        repo_name,
-        pr_number: comment_event.issue.number as i64,
-        comment_id: comment_event.comment.id.0 as i64,
-        requester_id,
-        requester_login: comment_event.comment.user.login.clone(),
-        body: body.clone(),
-        comment_created_at: comment_event.comment.created_at,
-    };
-
-    if let Err(e) = github_sender.send(task).await {
-        warn!(
-            "Failed to enqueue ProcessMergeCommand for PR #{}: {:?}",
-            comment_event.issue.number, e
-        );
-    } else {
-        debug!(
-            "Queued ProcessMergeCommand for PR #{} (comment {})",
-            comment_event.issue.number, comment_event.comment.id.0
-        );
-    }
-}
-
-async fn handle_github_installation(
-    installation_payload: payload::InstallationWebhookEventPayload,
-    installation: Option<EventInstallation>,
-    db_service: DbService,
-) {
-    use payload::InstallationWebhookEventAction as IWEA;
-
-    // Extract the full installation object from the event
-    let installation = match installation {
-        Some(EventInstallation::Full(inst)) => inst,
-        _ => {
-            warn!("Installation event missing full installation object");
-            return;
-        },
-    };
-
-    let installation_id = installation.id.0 as i64;
-    let account_type = &installation.account.r#type;
-    let account_login = &installation.account.login;
-
-    debug!(
-        "Received installation event: {:?} for installation {} ({})",
-        installation_payload.action, installation_id, account_login
-    );
-
-    match installation_payload.action {
-        IWEA::Created => {
-            // App was installed
-            if let Err(e) = db_service
-                .upsert_installation(installation_id, account_type, account_login)
+        WEP::IssueComment(comment_event) => {
+            comments::handle_github_issue_comment(*comment_event, repository_info, github_sender)
                 .await
-            {
-                warn!(
-                    "Failed to upsert installation {} for {}: {}",
-                    installation_id, account_login, e
-                );
-                return;
-            }
-
-            // Add all repositories that came with the installation
-            if let Some(repos) = installation_payload.repositories {
-                for repo in repos {
-                    let repo_id = repo.id.0 as i64;
-                    let repo_name = &repo.name;
-                    // Extract owner from full_name (format: "owner/repo")
-                    let repo_owner = repo.full_name.split('/').next().unwrap_or(account_login);
-
-                    if let Err(e) = db_service
-                        .upsert_installation_repository(
-                            installation_id,
-                            repo_id,
-                            repo_name,
-                            repo_owner,
-                        )
-                        .await
-                    {
-                        warn!(
-                            "Failed to add repository {} to installation {}: {}",
-                            repo_name, installation_id, e
-                        );
-                    } else {
-                        debug!(
-                            "Added repository {}/{} to installation {}",
-                            repo_owner, repo_name, installation_id
-                        );
-                    }
-                }
-            }
         },
-        IWEA::Deleted => {
-            // App was uninstalled
-            if let Err(e) = db_service.delete_installation(installation_id).await {
-                warn!("Failed to delete installation {}: {}", installation_id, e);
-            } else {
-                debug!("Deleted installation {}", installation_id);
-            }
-        },
-        IWEA::Suspend => {
-            // App was suspended
-            if let Err(e) = db_service.suspend_installation(installation_id).await {
-                warn!("Failed to suspend installation {}: {}", installation_id, e);
-            } else {
-                debug!("Suspended installation {}", installation_id);
-            }
-        },
-        IWEA::Unsuspend => {
-            // App was unsuspended
-            if let Err(e) = db_service.unsuspend_installation(installation_id).await {
-                warn!(
-                    "Failed to unsuspend installation {}: {}",
-                    installation_id, e
-                );
-            } else {
-                debug!("Unsuspended installation {}", installation_id);
-            }
-        },
-        action => {
-            debug!(
-                "Ignoring installation action {:?} for installation {}",
-                action, installation_id
-            );
-        },
-    }
-}
-
-async fn handle_github_installation_repositories(
-    installation_repos_payload: payload::InstallationRepositoriesWebhookEventPayload,
-    installation: Option<EventInstallation>,
-    db_service: DbService,
-) {
-    use payload::InstallationRepositoriesWebhookEventAction as IRWEA;
-
-    // Extract the installation object from the event
-    let installation = match installation {
-        Some(EventInstallation::Full(inst)) => inst,
-        Some(EventInstallation::Minimal(inst)) => {
-            // For installation_repositories events, we might get minimal installation
-            // We can still work with just the ID
-            debug!("Got minimal installation: {:?}", inst);
-            // We'll need the installation ID
-            let installation_id = inst.id.0 as i64;
-
-            debug!(
-                "Received installation_repositories event: {:?} for installation {}",
-                installation_repos_payload.action, installation_id
-            );
-
-            match installation_repos_payload.action {
-                IRWEA::Added => {
-                    // Repositories were added to the installation
-                    for repo in installation_repos_payload.repositories_added {
-                        let repo_id = repo.id.0 as i64;
-                        let repo_name = &repo.name;
-                        // Extract owner from full_name (format: "owner/repo")
-                        let repo_owner = repo.full_name.split('/').next().unwrap_or("");
-
-                        if let Err(e) = db_service
-                            .upsert_installation_repository(
-                                installation_id,
-                                repo_id,
-                                repo_name,
-                                repo_owner,
-                            )
-                            .await
-                        {
-                            warn!(
-                                "Failed to add repository {} to installation {}: {}",
-                                repo_name, installation_id, e
-                            );
-                        } else {
-                            debug!(
-                                "Added repository {}/{} to installation {}",
-                                repo_owner, repo_name, installation_id
-                            );
-                        }
-                    }
-                },
-                IRWEA::Removed => {
-                    // Repositories were removed from the installation
-                    for repo in installation_repos_payload.repositories_removed {
-                        let repo_id = repo.id.0 as i64;
-
-                        if let Err(e) = db_service
-                            .delete_installation_repository(installation_id, repo_id)
-                            .await
-                        {
-                            warn!(
-                                "Failed to remove repository {} from installation {}: {}",
-                                repo_id, installation_id, e
-                            );
-                        } else {
-                            debug!(
-                                "Removed repository {} from installation {}",
-                                repo_id, installation_id
-                            );
-                        }
-                    }
-                },
-                action => {
-                    debug!(
-                        "Ignoring installation_repositories action {:?} for installation {}",
-                        action, installation_id
-                    );
-                },
-            }
-            return;
-        },
-        None => {
-            warn!("Installation_repositories event missing installation object");
-            return;
-        },
-    };
-
-    let installation_id = installation.id.0 as i64;
-
-    debug!(
-        "Received installation_repositories event: {:?} for installation {}",
-        installation_repos_payload.action, installation_id
-    );
-
-    match installation_repos_payload.action {
-        IRWEA::Added => {
-            // Repositories were added to the installation
-            for repo in installation_repos_payload.repositories_added {
-                let repo_id = repo.id.0 as i64;
-                let repo_name = &repo.name;
-                // Extract owner from full_name (format: "owner/repo")
-                let repo_owner = repo
-                    .full_name
-                    .split('/')
-                    .next()
-                    .unwrap_or(&installation.account.login);
-
-                if let Err(e) = db_service
-                    .upsert_installation_repository(installation_id, repo_id, repo_name, repo_owner)
-                    .await
-                {
-                    warn!(
-                        "Failed to add repository {} to installation {}: {}",
-                        repo_name, installation_id, e
-                    );
-                } else {
-                    debug!(
-                        "Added repository {}/{} to installation {}",
-                        repo_owner, repo_name, installation_id
-                    );
-                }
-            }
-        },
-        IRWEA::Removed => {
-            // Repositories were removed from the installation
-            for repo in installation_repos_payload.repositories_removed {
-                let repo_id = repo.id.0 as i64;
-
-                if let Err(e) = db_service
-                    .delete_installation_repository(installation_id, repo_id)
-                    .await
-                {
-                    warn!(
-                        "Failed to remove repository {} from installation {}: {}",
-                        repo_id, installation_id, e
-                    );
-                } else {
-                    debug!(
-                        "Removed repository {} from installation {}",
-                        repo_id, installation_id
-                    );
-                }
-            }
-        },
-        action => {
-            debug!(
-                "Ignoring installation_repositories action {:?} for installation {}",
-                action, installation_id
-            );
-        },
+        // We probably don't want to react to every push
+        // WEP::Push(pr) => handle_github_push(*pr).await,
+        _ => (),
     }
 }
